SAS-92 #comment added pre-commit support with some tools
