+++ Makefile
@@ -4,6 +4,13 @@ build-dev:
 build-test:
 	@sudo docker build --pull --rm -f "Dockerfile.test" -t test .
 
+clean:
+	@rm -rf */__pycache__ */*/__pycache__ src/db/*.sqlite
+
+delete-containers:
+	@sudo docker stop devcontainer || true && sudo docker rm devcontainer || true
+	@sudo docker stop testcontainer || true && sudo docker rm testcontainer || true
+
 mkenv:
 	@./configs/setenv.sh
 
@@ -18,10 +25,3 @@ run-local:
 
 test: delete-containers build-test
 	@sudo docker run -it --name testcontainer -p 8000:8000 test
-	
-delete-containers:
-	@sudo docker stop devcontainer || true && sudo docker rm devcontainer || true
-	@sudo docker stop testcontainer || true && sudo docker rm testcontainer || true
-
-clean:
-	@rm -rf */__pycache__ */*/__pycache__ src/db/*.sqlite
