+++ configs/requirements-test.txt
@@ -3,4 +3,3 @@ pytest
 flake8
 pep8
 coverage
-autopep8
